Temp fix for JSON parsing
I tracked down and changed the arrow function that was breaking JSON
parsing. Replacing it with a nameless (what I have a habit of calling
"lambda") function fixed the issue. I'm committing this so we can
continue working on other parts of the page; we can fix this particular
issue later. Just a total guess: We might have to look closer at how
arrow functions handle "this".