Fixed arrow function
Put the arrow function back in, so eslint can now be happy. It was a
scope issue for "this" like I thought. I just sent it in as a private
variable called "._that", and everything worked fine. We'll likely have
to do the same for any arrow function using "this" (or any other
volatile global,